requirements.txt
numpy
pandas
matplotlib
scipy
SQLAlchemy
Jupyter
openpyxl
xlrd
python-dotenv
colorlog
pyupdater

quart
poetry
ttkbootstrap
pillow
PySide6
qt-material
watchdog
